OpenCL Kernel #2
Added the ability to call out to kernels and standardized a standard
input for images along as decoupled the image from the actual image
processing done by kernels. Also added a bit of code to optimize work
groups. Fixed a minor error in the scroll bars in where they would over
scroll the image by 10 pixels due to leftover code.