Acceler Format
Ability to save images in the Acceler format (.ACL) which allows for
layers and composite images to be saved. .ACL also supports metadata to
be stored about the layers making room to save more complex effects.
Visibility is stored in metadata along with Opacity which has yet to be
implemented.